third_party/blink/web_tests/external/wpt/css/css-text/parsing/hanging-punctuation-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['hanging-punctuation'] = "force-end" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['hanging-punctuation'] = "first force-end" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['hanging-punctuation'] = "force-end last" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['hanging-punctuation'] = "first last force-end" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['hanging-punctuation'] = "force-end first last" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['hanging-punctuation'] = "last first force-end" should set the property value
  assert_not_equals: property should be set got disallowed value ""
Harness: the test ran to completion.
